Add A Basic Fancy Readme
This commit adds a basic (fancy) readme to the project. The readme this
commit provides is in markdown, and primarly serves as a template for
further readme development. On a side note, this commit also adds a
basic stand-in project logo that was made using the toilet command and
paint.net. As such, there is no need to worry about rights to the logo.

Since the readme is the most important part of a project, this commit is
the most important commit.