Preliminary Documentation
This is a test of how to document the code. I am not on my mac right now
so I'm not entirely certain if this is 100% correct or if it has the
correct formatting. Regardless, this can serve as a starting point in
developing the method by which we want to document the code.